packages/clipper-core/pyproject.toml
[project]
name = "clipper-core"
version = "0.1.0"
description = "Auto Clipper Core - Pure Python library for video clipping"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
]

[project.optional-dependencies]
full = [
    "openai-whisper>=20231117",
    "google-generativeai>=0.3.0",
    "groq>=0.4.0",
    "openai>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["clipper_core"]
